+++ split/test/src/split_test.cpp
@@ -81,7 +81,7 @@ static void compare(const std::vector<std::string>& file_names, bool do_byte_swa
         vrt_header prev_header;
         vrt_fields prev_fields;
 
-        size_t i;
+        uint64_t i;
         for (i = 0;; ++i) {
             // Make space for header
             if (buf.size() < VRT_WORDS_HEADER) {
